move from directive to component
fix aot build
