+++ .angular-cli.json
@@ -6,7 +6,7 @@
   "apps": [
     {
       "root": "demo",
-      "outDir": "dist",
+      "outDir": "demo-dist",
       "assets": [
         "assets",
         "favicon.ico"
