+++ demo/app/app.component.html
@@ -1,5 +1,5 @@
 <div style="text-align:center; position: relative">
-  <div [ngxLoader]="{busy: busySelected, backdrop: activeBackdrop, customClass:'myclass'}"></div>
+  <ngx-loader [loader]="{subject: subjectSelected, backdrop: activeBackdrop, customClass:'myclass'}"></ngx-loader>
   <h1>
     Welcome to this demo!
   </h1>
