+++ demo/app/app.component.ts
@@ -1,5 +1,4 @@
-import {Component, OnInit} from '@angular/core';
-import {Subscription} from 'rxjs/Subscription';
+import {Component} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
 import 'rxjs/add/operator/delay';
@@ -9,12 +8,12 @@ import 'rxjs/add/operator/delay';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.sass'],
 })
-export class AppComponent{
-  busySelected: any;
-  activeBackdrop: boolean = false;
+export class AppComponent {
+  subjectSelected: any;
+  activeBackdrop = false;
 
   returnPromise() {
-    this.busySelected = new Promise((resolve) => {
+    this.subjectSelected = new Promise(resolve => {
       setTimeout(() => {
         resolve();
         }, 2000);
@@ -22,6 +21,6 @@ export class AppComponent{
   }
 
   returnSubscription() {
-    this.busySelected = Observable.of('string').delay(2000).subscribe();
+    this.subjectSelected = Observable.of('string').delay(2000).subscribe();
   }
 }
