+++ gulpfile.js
@@ -183,8 +183,8 @@ gulp.task('compile', function () {
     'copy:source',
     'inline-resources',
     'ngc',
-    'rollup:fesm',
-    'rollup:umd',
+    // 'rollup:fesm',
+    // 'rollup:umd',
     'copy:build',
     'copy:manifest',
     'copy:readme',
