+++ src/index.ts
@@ -1,5 +1,4 @@
-export {NgxLoaderModule} from './ngx-loader.module';
-export {NgxLoaderDirective} from './ngx-loader.directive';
-export {NgxLoaderComponent} from './ngx-loader.component';
-export {NgxLoaderBackdropComponent} from './ngx-loader-backdrop.component';
-export {NgxLoaderConfig} from './ngx-loader.config';
+export * from './ngx-loader.module';
+export * from './ngx-loader.component';
+export * from './ngx-loader-backdrop.component';
+export * from './ngx-loader.config';
