+++ src/ngx-loader.config.ts
@@ -6,8 +6,8 @@ export class NgxLoaderConfig implements INgxLoaderConfig {
   customClass: string;
 
   constructor(config: INgxLoaderConfig = {}) {
-    for (const option in BUSY_CONFIG_DEFAULTS) {
-      this[option] = config[option] != null ? config[option] : BUSY_CONFIG_DEFAULTS[option];
+    for (const option in SUBJECT_CONFIG_DEFAULTS) {
+      this[option] = config[option] != null ? config[option] : SUBJECT_CONFIG_DEFAULTS[option];
     }
   }
 }
@@ -16,13 +16,13 @@ export interface INgxLoaderConfig {
   template?: string;
   backdrop?: boolean;
   customClass?: string;
-  busy?: Promise<any> | Subscription | Array<Promise<any> | Subscription>;
+  subject?: Promise<any> | Subscription | Array<Promise<any> | Subscription>;
 }
 
-export const BUSY_CONFIG_DEFAULTS = {
+export const SUBJECT_CONFIG_DEFAULTS = {
   template: `
-        <div class="ng-busy-default-wrapper"></div>
+        <div class="ng-subject-default-wrapper"></div>
     `,
   backdrop: false,
-  customClass: 'ng-busy'
+  customClass: 'ng-subject'
 };
