+++ src/tsconfig.es5.json
@@ -8,23 +8,20 @@
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "moduleResolution": "node",
-    "outDir": "../build",
+    "outDir": "../dist",
     "rootDir": ".",
     "lib": [
+      "es6",
       "es2015",
       "dom"
     ],
-    "skipLibCheck": true,
-    "types": []
+    "noImplicitAny": false,
+    "sourceMap": true,
+    "mapRoot": "",
+    "skipLibCheck": true
   },
   "angularCompilerOptions": {
-    "annotateForClosureCompiler": true,
-    "strictMetadataEmit": true,
     "skipTemplateCodegen": true,
-    "flatModuleOutFile": "ngx-loader.js",
-    "flatModuleId": "ngx-loader"
-  },
-  "files": [
-    "./index.ts"
-  ]
+    "strictMetadataEmit": true
+  }
 }
