+++ tsconfig.json
@@ -4,7 +4,6 @@
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
     "moduleResolution": "node",
-    "rootDir": "./src",
     "lib": [
       "es2015",
       "dom"
